Add extension to create codegen task
