+++ codegen/server/build.gradle.kts
@@ -11,27 +11,16 @@ dependencies {
     implementation(project(":server-core"))
 }
 
-// Execute building of Java classes using an executable class
-// These classes will then be used by integration tests and benchmarks
-val generatedSrcDir = layout.buildDirectory.dir("generated-src").get()
-val generateSrcTask = tasks.register<JavaExec>("generateSources") {
-    delete(generatedSrcDir)
-    dependsOn("test")
-    classpath = sourceSets["test"].runtimeClasspath + sourceSets["test"].output + sourceSets["it"].resources.sourceDirectories
-    mainClass = "software.amazon.smithy.java.codegen.server.TestServerJavaCodegenRunner"
-    environment("service", "smithy.java.codegen.server.test#TestService")
-    environment("namespace", "smithy.java.codegen.server.test")
-    environment("output", generatedSrcDir)
-}
-
 tasks {
-    test {
-        finalizedBy("integ")
-    }
+    val generateSrcTask = addGenerateSrcsTask("software.amazon.smithy.java.codegen.server.TestServerJavaCodegenRunner")
+
     integ {
         dependsOn(generateSrcTask)
     }
     compileItJava {
         dependsOn(generateSrcTask)
+    }
+    test {
+        finalizedBy(integ)
     }
 }
