+++ codegen/server/src/test/java/software/amazon/smithy/java/codegen/server/TestServerJavaCodegenRunner.java
@@ -30,8 +30,8 @@ public final class TestServerJavaCodegenRunner {
             .fileManifest(FileManifest.create(Paths.get(System.getenv("output"))))
             .settings(
                 ObjectNode.builder()
-                    .withMember("service", System.getenv("service"))
-                    .withMember("namespace", System.getenv("namespace"))
+                    .withMember("service", "smithy.java.codegen.server.test#TestService")
+                    .withMember("namespace", "smithy.java.codegen.server.test")
                     .build()
             )
             .model(model)
